requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
cloud-sql-python-connector[pg8000]
httpx==0.27.2
anthropic==0.40.0
PyYAML==6.0.2
pytest==8.3.5
openai==1.68.2
